Make use of syscalls to NTDLL when using VirtualProtect + prepare VersioningStruct
The newest RS employs a version of VMProtect where  pretty much all relevant parts of .text section are write protected, and it also hooks VirtualProtect and similar methods to return ACCESS_DENIED status on every call.

Using syscalls to directly call the methods from NTDLL.dll is a way to avoid that.

Also since we want to support both the old and new RS versions, we introduce the VersioningStruct to keep the offsets for old and new versions.